IGNITE-14917 Configuration modules internal packages renamed №2.
