+++ modules/api/src/main/java/org/apache/ignite/configuration/schemas/network/NetworkConfigurationSchema.java
@@ -18,10 +18,10 @@
 package org.apache.ignite.configuration.schemas.network;
 
 import org.apache.ignite.configuration.annotation.ConfigurationRoot;
+import org.apache.ignite.configuration.annotation.ConfigurationType;
 import org.apache.ignite.configuration.annotation.Value;
 import org.apache.ignite.configuration.validation.Max;
 import org.apache.ignite.configuration.validation.Min;
-import org.apache.ignite.internal.configuration.storage.ConfigurationType;
 
 /**
  * Configuration schema for network endpoint subtree.
