+++ modules/api/src/main/java/org/apache/ignite/configuration/schemas/runner/NodeConfigurationSchema.java
@@ -18,8 +18,8 @@
 package org.apache.ignite.configuration.schemas.runner;
 
 import org.apache.ignite.configuration.annotation.ConfigurationRoot;
+import org.apache.ignite.configuration.annotation.ConfigurationType;
 import org.apache.ignite.configuration.annotation.Value;
-import org.apache.ignite.internal.configuration.storage.ConfigurationType;
 
 /**
  * Local node configuration schema.
