+++ modules/api/src/main/java/org/apache/ignite/configuration/schemas/table/TablesConfigurationSchema.java
@@ -18,8 +18,8 @@
 package org.apache.ignite.configuration.schemas.table;
 
 import org.apache.ignite.configuration.annotation.ConfigurationRoot;
+import org.apache.ignite.configuration.annotation.ConfigurationType;
 import org.apache.ignite.configuration.annotation.NamedConfigValue;
-import org.apache.ignite.internal.configuration.storage.ConfigurationType;
 
 /**
  * Tables configuration schema.
