+++ modules/configuration-annotation-processor/src/test/java/org/apache/ignite/configuration/sample/NetworkConfigurationSchema.java
@@ -19,7 +19,7 @@ package org.apache.ignite.configuration.sample;
 
 import org.apache.ignite.configuration.annotation.ConfigValue;
 import org.apache.ignite.configuration.annotation.ConfigurationRoot;
-import org.apache.ignite.internal.configuration.storage.ConfigurationType;
+import org.apache.ignite.configuration.annotation.ConfigurationType;
 
 /**
  * Test network configuration schema.
