+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/NamedConfigurationTree.java
@@ -18,8 +18,6 @@
 package org.apache.ignite.configuration;
 
 import org.apache.ignite.configuration.notifications.ConfigurationNamedListListener;
-import org.apache.ignite.internal.configuration.tree.NamedListChange;
-import org.apache.ignite.internal.configuration.tree.NamedListView;
 
 /**
  * Configuration tree representing arbitrary set of named underlying configuration tree of the same type.
