+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/NamedListChange.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.configuration.tree;
+package org.apache.ignite.configuration;
 
 import java.util.function.Consumer;
