+++ modules/configuration-api/src/main/java/org/apache/ignite/configuration/annotation/ConfigurationType.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.configuration.storage;
+package org.apache.ignite.configuration.annotation;
 
 /**
  * Type of configuration storage.
